Generator experiment
Not sure if I like this idea but it was something that seemed
interesting at one point in time. This allows generation of a
a wrapper around the set implementations that asserts concrete types for
you. This utility could be run with "go generate" to build support at
compile time. Maybe I'll come back to this at some point in time.